+++ changelog.md
@@ -1,8 +1,8 @@
 # Changelog
 
-All notable changes to `LaravelUtilities` will be documented in this file.
+All notable changes to `laravel-utilities` will be documented in this file.
 
-## Version 1.0
+## Version 1.0.0
 
 ### Added
-- Everything
+- Ascii validation rule